[gen-utils] Adds a few string casing utils.
