+++ packages/labs/gen-wrapper-react/src/index.ts
@@ -12,7 +12,8 @@ import {
   Package,
   PackageJson,
 } from '@lit-labs/analyzer/lib/model.js';
-import {javascript, FileTree} from '@lit-labs/gen-utils/lib/file-utils.js';
+import {FileTree} from '@lit-labs/gen-utils/lib/file-utils.js';
+import {javascript, kabobToOnEvent} from '@lit-labs/gen-utils/lib/str-utils.js';
 
 export const generateReactWrapper = async (
   analysis: Package
@@ -152,11 +153,6 @@ ${elements.map((element) => wrapperTemplate(element))}
 // TODO(kschaaf): Should this be configurable?
 const packageNameToReactPackageName = (pkgName: string) => `${pkgName}-react`;
 
-const eventNameToCallbackName = (eventName: string) =>
-  'on' +
-  eventName[0].toUpperCase() +
-  eventName.slice(1).replace(/-[a-z]/g, (m) => m[1].toUpperCase());
-
 const wrapperTemplate = ({name, tagname, events}: LitElementDeclaration) => {
   return javascript`
 export const ${name} = createComponent(
@@ -166,7 +162,7 @@ export const ${name} = createComponent(
   {
     ${Array.from(events.keys()).map(
       (eventName) => javascript`
-    ${eventNameToCallbackName(eventName)}: '${
+    ${kabobToOnEvent(eventName)}: '${
         // TODO(kschaaf): add cast to `as EventName<EVENT_TYPE>` once the
         // analyzer reports the event type correctly (currently we have the
         // type string without an AST reference to get its import, etc.)
@@ -175,6 +171,6 @@ export const ${name} = createComponent(
       }',`
     )}
   }
-);    
+);
 `;
 };
